Moved auxiliary functions to utils
